Make SLARGState non-serializable
The serialization that was implemented would never really work,
because it references CFAEdge instances and would serialize the whole
CFA together with a SLARGState.
But when we (de-)serialize abstract states,
we do not want to serialize the CFA,
we want to add references to the then-present instance of the CFA
during deserialization.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@47083 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c