Remove Serializable from CFA nodes, edges and metadata
As discussed in #1214, serialization of CFAs is not used in practice,
it is likely broken and it will probably never be used.
But maintaining it is tricky due to the common problems with Java
serialization, so let's remove it.

We want to keep serializability of Type instances for now,
because some abstract states reference them,
and we want to be able to serialize those.
Unfortunately, CArrayType stores an arbitrary CExpression,
so all C expressions and transitively also almost all C AST nodes
need to remain serializable as well.
And because of how Java serialization works,
all super types needs to remain serializable
and all subtypes of those as well.
So we end up with all AST nodes remaining serializable,
which is unfortunately much more than desired.

Fixes #1214.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@47085 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c