Fix Junit tests.
The new solver OPENSMT uses a formula representation "a=1" instead of "1=a",
due to solver-internal formula-rewriting,
and the set of atoms after splitting atoms into inequalities missed "1<=a" and provided "a<=1".
Overall, only the first of the inqualities was ever considered for further processing.
It is unclear whether this was a bug or not.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/update-javasmt-to-v5.0.0@47165 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c