Create ValueAnalysisResultToLoopInvariants instance only if enabled
There are several reasons for this.

First, this class requires an SMT solver, but our default SMT solver
is not available on all platforms, so if the user deliberately chooses a
configuration of the value analysis that does not require an SMT solver,
the code here would break running CPAchecker on non-Linux platforms.

Second, we can optimize and save the initialization effort
(the class even does some preliminary work in the constructor).

Third, there are several different exceptions being produced
during the initialization of this class (cf. BuildBot),
which currently breaks most configurations of CPAchecker.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@47298 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c