Update SoSy-Lab Java Common Lib
This brings as feature that deprecated names of config options
are now also documented in doc/ConfigurationOptions.txt.
This is useful for users who might see a deprecated name being used
somewhere and want to look it up.

Due to a new compilation check by the library
we also notice that the deprecated name maxPreciseStrlenSize
was not actually usable due to a bug in the way the config options
were declared (deprecatedName works only in combination with
deprecatedPrefix).
As this does not seem to have created problems,
we can just delete this deprecated name.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@47326 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c