Clean up switch in CFABuilder
There are lots of empty cases that just fall through to the default
case, which is the "not implemented yet" case.
With the current style it can easily happen that someone implements one
of these cases later, and then all the cases above would also delegate
to this case instead of the "not implemented yet" case.
It is safer to delete all cases and let them directly go to the default
case.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@47364 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c