Remove some unnecessary or confusing "fall through" comments.
Such comments are intended for cases where one non-empty case
on purpose continues with the next case, because this is a highly
unusual and often confusing situation.
Fall through of empty cases is common and does not warrant a comment.
Such comments would also remain after a conversion to arrow labels,
where they then are just confusing
(cf. FaultLocalizationWithTraceFormula).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@47365 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c