Update google-java-format to 1.23.0
Please update your IDE plugins as well,
because this brings a change in how comments for case statements are
formatted. A comment before a case typically applies to that case,
not the previous case, so it should be indented accordingly.
A good example are fall-through markers,
which connect two cases and should thus be on the same level as the
cases and not hidden behind more indentation.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@47366 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c