Support BigInteger in BDDBooleanExpressionVisitor
We always expect values 0 or 1, but we still must be precise about large integers
that do not fit into a Java long. If we just truncated these values, we might get
to 0 or 1 by chance, and produce a hard-to-find bug.
To avoid this, we are precise about large integers and throw the same exception
that is thrown for other values != 0,1

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/make-cenumerator-use-biginteger@47394 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c