Move ever-growing cache of ASTTypeConverter from static field to ParserContext.
The cache of IType/CTypes can grow very big (multiple hundred MB).
As a static field, this cache is never cleared, but kept in memory.
To avoid this, we want to give the cache a lifetime that matches
the time that it is actually useful.
By associating the cache to the ParserContext, this is fulfilled:
The parser context lives as long as the relevant CFA creation,
so it can contain caches that are relevant for the runtime of the CFA
creation. These are garbage-collected afterwards, together with the
ParserContext.

There is one issue: The GlobalScope is supposed to hold information
about already-transformed C types, so it should not know anything about
the ParseContext or eclipse-specific types.
This principle is broken: GlobalScope resolves elaborate types to the actual types when they are registered in the scope.
In this process, it also tries to optimize the cache in ASTTypeConverter so that C types
are already converted to the resolved (non-elaborate) types.
Our hypothesis is that this is not necessary, so we remove
the cache optimization to keep the class architecture clean.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/make-cpachecker-usable-as-library@47700 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c