Output information when exporting witnesses about wether the translation from ARG states to ExpressionTrees was possible or not
In particular this is relevant for the predicate analysis where the SMT formulas cannot always be mapped back to an ExpressionTree.
This also allows the user to get some information about the guarantees the witness provides i.e. the invariants contained in it will likely not be inductive nor safe.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@47942 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c