Change several options in --dataFlowAnalysis
So far, this config just used the "invariant generation" config.
But for generating and exporting invariants we need some options
that we do not need if we just want to analyze program safety.
So we split this into a separate config file.
With this commit, --dataFlowAnalysis no longer uses:
- the ARGCPA
- cpa.automaton.breakOnTargetState = 0
- cpa.composite.aggregateBasicBlocks = false

cf.
https://gitlab.com/sosy-lab/software/cpachecker/-/issues/1157#note_2118965184

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1157-dataflow-configs@47975 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c