Choose specific JVM garbage collectors for CPAchecker
The default garbage collector of the JVM is the G1GC,
but for our use case (where in particular pause time and latency are not
so important) other garbage collectors can be more efficient.
A study was made by Tobias Maget (linked in the code)
across different analyses of CPAchecker and the SV-Benchmarks task set,
and this commit implements the resulting recommendations.

We distinguish between two main use cases:

- Regular users just want to get the result of CPAchecker as fast as
  possible.
- Researchers want to benchmark CPAchecker configurations measuring
  CPU time, and value deterministic performance.

For each use case there is a separate GC recommendation.
We choose between them based on whether "--benchmark" option is used.
If the user passes a parameter that selects a specific GC,
we make sure to use that and not our default GC.

Fixes #1154, except for Windows.
For more details, cf. the linked study.