LiveVariables: Add support for non-returning functions like `exit()`
LiveVariable analysis starts at the return node of a function and then works backwards to calculate the live variables for all locations in the function. When non-returning functions like `exit()` are called this approach won't work as such calls add additional end nodes to the functions. Because of this code before the non-return function can't be reached by backtracking from the return node.

We solve this issue by calculating all final nodes that can be reached from the function entry node and then including all of them in the initial state for the LiveVariable analysis.