Guard reached-set sharing in ParallelAlgorithm behind ::supply-reached
Contrary to what would be intuitive, the ::supply-reached annotation
so far decided only whether the reached set would be shared after the
analysis finishes. But for all analyses with a ReachedSetUpdateListener
(in practice the CEGARAlgorithm) the reached-set sharing would be always
active. This is not intuitive and likely unwanted.
So now even analyses with a ReachedSetUpdateListener must have the
::supply-reached annotation to enable reached-set sharing.