Check for missing loop information upfront in PredicateCPA
If we configure abstractions to happen at loop heads,
the PredicateCPA needs the loop information,
which is not always present (cf. #307).
So far it threw an IllegalStateException in this case
because we consider the missing loop information a bug
that needs to be fixed.
But this hurts in a portfolio approach,
where we want to continue with other analyses if the PredicateCPA
cannot work. So let's throw a CPAException from the constructor instead,
like for example LoopBoundCPA does as well.