Simplify detection of version number of development versions
So far we only computed an exact version number (with commit hash)
when we built a JAR file, not on standard compilation.
This was done in order to not impose the cost of running "svn stat"
every time.
Now that we only use git and since storage is faster than 10 years ago,
we can likely remove that optimization and just use the exact version
all the time.