Add a script for generating tables from BuildBot results
This can be used to quickly create and show a table
that compares two or more specific BuildBot jobs
by just passing the ids of the builder and the jobs.

I find it useful for catching up with BuildBot results
after some time of not looking at them,
instead of going through all the mails and tables one-by-one.