Tweak instructions for CPAchecker git checkouts
There are two commands that can be used to configure git locally
such that it handles the CPAchecker repository better.
One fixes a gap in old parts of the history,
and the other improves git blame with respect to reformatting commits.
This commit describes them in the documentation
and adds the necessary list of reformatting commits.
These commits were selecting by searching for commits
with "reformat" in the message and many changes.