Remove MBean from PredicateCPA's AbstractionManager
MBeans are a management interface that can be exposed to clients
outside of the JVM and was added as a way to monitor the number of
predicates. However, it causes trouble if two instances of the CPA are
active at the same time, and we never really used it.
So let's just remove it.

Fixes #1294.