Add Unit Test for LemmaEntry
Adds a unit test to test that a LemmaEntry fields 'value' and 'format'
get instantiated correctly from a 'witness.yml' file.
In this state the test throws an error.
To fix the error, either a custom deserializer has to be implemented for
'LemmaEntry' or the class hierarchy has to be restructured.