Fix endpoint path in webclient tool revision request.
Adjusted the endpoint path in `_request_tool_revision` to use `version_string` instead of `revision_string`. This aligns the request with the correct API endpoint ensuring backwards compatibility with the old versions of this script.

I had changed the name because I find confusing to have two variables using the word `version`. I still find it is so, so, I let a clarify comment about:
- _request_tool_revision asks the API for the revision reference (a full commit hash)
- request_tool_version asks the API for the version number that the tool generates itself. How this version number is obtained, is defined in the configurable option `getToolVerion` in the class `WebClientConfiguration` on the BenchCloud side.