[Witness Config Handling] Correctly add the invariants to predicate analysis for validating no-overflow correctness witnesses
Currently the witness was being ignored, since the option `witness.validation.correctness.isa` was in the sub-config for validating no-overflow witnesses, which was not being read till later. In principle this meant that the predicate analysis was ignoring the witness.
This commit now does the full config processing to get the option and then continues as before.