Refactoring: control-flow simplification in ARGUtils (IV)
In producePathAutomatonWithLoops there is a big "if-else" chain
where some cases had redundant continue statements.
To clean this up, also move the comments inside the case that they
describe. Otherwise we would have cases that would be fully empty.

This code was flagged by the RedundantControlFlow check of Error Prone.

Part of #1307.