Use a single instance of the merge operator in the PredicateCPA
The use of separate instances of operators was
intended to allow multi-threaded usage of the PredicateCPA.
But this did not work out according to
https://gitlab.com/sosy-lab/software/cpachecker/-/commit/70298e2de1b4d1a258e57e2ba34c07636ec6e72f#note_647467995
and will likely never work because solvers are not thread safe.

Having a single instance that is instantiated from the PredicateCPA
constructor has the following advantages:
- The merge code together with its options and statistics can be inside
  the same class.
- The options for merge are only considered if this merge operator is
  actually used, and users get a warning about unused options otherwise
  (as desired, but currently this is not the case).
- The statistics code can correctly print the statistics for merge
  depending on whether the merge operator was configured to be used,
  and not depending on whether it happened to be called.

This reverts parts of 70298e2d.