Remove class WeakeningOptions
The options for the various classes that are part of the
inductive-weakening component should be in their respective classes,
as usual. This is better encapsulation.
Furthermore, right now a user does not get a warning about an unused
option if they specify an option that is part of an unused
sub-component, because all options are parsed always.
With this commit this is improved and warnings about unused options
are produced as they should be.