Use a single instance of the transfer relation in the PredicateCPA
The use of separate instances of operators was
intended to allow multi-threaded usage of the PredicateCPA.
But this did not work out according to
https://gitlab.com/sosy-lab/software/cpachecker/-/commit/70298e2de1b4d1a258e57e2ba34c07636ec6e72f#note_647467995
and will likely never work because solvers are not thread safe.

This change moves the options and statistics of PredicateTransferRelation
back into this class, which is better for code coherency.
Furthermore, they were previously even stored inconsistently
(some statistics were copied into PredicateTransferRelation,
some were only stored in PredicateStatistics).

This also allows us to get rid of the classes PredicateStatistics
(which did not contain statistics about predicates,
and is named confusingly similar to the class PredicateCPAStatistics)
and PredicateCpaOptions
(which did not really contain the options of the PredicateCPA,
only those of the transfer relation).

This reverts parts of 5124b100 and 69726590.