Use 'env bash' in scripts instead of hardcoding '/bin/bash'
The use of /bin/bash executes scripts with the default bash of the user's OS,
but the user may have installed other versions of bash and put them in their path
before bin/bash .
A common case of this is on macOS: The bash version shipped with macOS is
super-old (3.2.57(1)-release), so many users install a newer bash
through homebrew or other means (my current version: 5.2.37(1)-release).

This change is not just a nitpick: The bash shipped with macOS
is so old that it does not understand some of the syntax used in our scripts.
/bin/cpachecker currently fails with this error when executed with /bin/bash:

bin/cpachecker: line 73: syntax error near unexpected token `;'
bin/cpachecker: line 73: `       ;&'

This commit fixes this, as long as users actually use a newer bash.