Enable some more static-analysis checks of ECJ
And do not clutter the output on the command line with "info" messages.
These info messages are nice hints when looking at the respective line,
but are too many and with too many false alarms in order to look at them
in general and try to fix all of them.