Parse LemmaEntry as ExpressionTree
Implements a method "parseLemma" that returns an ExpressionTree from a
LemmaEntry so in the next step we can generate a BooleanFormula
representing the Lemma.

Parse LemmaEntry as ExpressionTree

Implements a method "parseLemma" that returns an ExpressionTree from a
LemmaEntry so in the next step we can generate a BooleanFormula
representing the Lemma.