Fix handling of function calls with name inside parentheses
These are cases such as "(f)()". So far we rewrite them into "(*(&f))()",
which is unnecessary and unwanted.
The FunctionPointerCPA cannot handle this, for example.
But "(f)()" is the same as "f()", we just don't detect this so far.
This commit fixes that (for an arbitrary number of nested parentheses
around the function name).

While we are at it, apply the same fix for cases where we match "&foo",
such that "&(foo)" now works just like "&foo".

As a test case we can just reuse one of our existing programs.

Fixes #1321.