Parse Declarations from witness
A lemma set entry in the yaml witness contains a list of declarations.
The declarations are either function declarations or variable
declarations. Function declarations have a return type, a function name
and a possibly empty list of parameters. Variable declarations consist
of a type and the variable name.
The declarations within a lemma set entry must be fresh and can not
occur somewhere else in the source code.
Declarations are parsed into CDeclaration objects. These CDeclarations
are added to the program scope.