Correctly parse the calculation type for ACSL statements
To parse ACSL statements into CExpressions, function calls in the
statement are replaced with temporary variables. The function call and
the remaining statement are parsed into CExpressions seperately. The
function call expression is then back inserted into the resulting
CBinaryExpression.
To get the correct calculation type of the binary expression, the
declarations of the temporary variables have to be added to the
CProgramScope temporarily and need to be removed from the scope again
after the expression could be parsed correctly.