Add copy constructor for CProgramScope
Using the copy contsructor the parsing of ACSL expressions creates a new
temporary scope that is a copy of the C program scope instead of
operating on a reference of the actual program scope. This way the
actual program scope is not changed by the parsing of ACSL function
calls.