Parse predicates containing ACSL expressions
The predicates provided in the witness file can reference lemmas and can
contain ACSL function calls. These function calls need to be treated as
ACSL statements while parsing.

The changes within this commit are a quick fix to get the parsing of
predicates running with function calls. It needs to be reimplemented
properly and should not be used anywhere else.