Refactoring: use pattern-matching instanceof with manual name choices
This introduces pattern-matching instanceof in cases
where Error Prone could not find a non-clashing name
and so the variable names had to be chosen manually.

This commit changes only cases where the instanceof check is negated.
Some cases have a method call in the expression
that is now evaluated only once, but it was checked manually
that it is free of side effects.