Refactoring: use pattern-matching instanceof (PMA)
This introduces pattern-matching instanceof for cases that
- have a positive (non-negated) instanceof check,
- where a method call exists inside the instanceof and cast expressions,
  and
- where the value is assigned to a variable after the cast.

This could in theory change the semantics of the code
if the method has a side effect,
but this is quite unlikely and code like
if (f(foo) instanceof Bar) { ((f(foo)) Bar).bar(); }
should be changed anyway if "f" has a side effect
because this is just too confusing.
While doing some spot checks on the commit
I found only trivial getters and similar method calls.

The variable name in the pattern-matching instanceof
is taken from the assignment that is now removed,
so all names are the same as before.

Commit created with Google Error Prone.