Convert switches with classic case labels to arrow cases with Error Prone
The new syntax does not allow fallbacks and variable declarations
that leak outside of the case, so it is more clear.
This is one step more towards getting rid of classic case labels completely.
Now mostly switches with a fall through between non-empty cases
or with a fall through into a default case are missing to be converted.

Commit generated with Google Error Prone 2.38.0.