Refactor a few switch statements to avoid fall through into default
Using the same code for two or more cases is supported
by the new switch syntax with arrow labels,
but a case and the default case can not be combined.
So if we want to refactor such switch statements
we need to duplicate the code.
This is mostly not bad because it is usually only a single throw,
and after the refactoring we can often actually remove the default case.

This commit does not handle all such cases, only a few.