Refactoring: avoid fall through by moving code
Because we check the full function name anyway,
the check for getKind() == BV_EXTRACT is redundant
and the code can be in the default case as well,
as long as no BV_EXTRACT case exists.
This is likely a good enough solution until the code is rewritten,
which seems necessary anyway.