Remove default clauses of switch expressions that are now redundant
This is for cases where a switch expression for an enum exists,
and all enum constants are handled in specific cases.
Removing the default clause will make it a compilation failure
instead of a runtime error if the respective enum is extended.

Commit automatically generated with the UnnecessaryDefaultInEnumSwitch
check of Google Error Prone (which is not enabled by default).
Afterwards two cases of throws clauses could be cleaned up.

One can easily check that all removed default clauses are actually in
switch expressions (and not switch statements),
because the former have "};" on the line after the default clause,
whereas the latter would have only "}".

Part of #1339.