Split logical operation (unsafe) taint tests into separate files
Separated (unsafe) taint analysis tests for logical operations `&&` and `||` into individual files for better clarity and granularity.

Here there is still an issue to resolve, since the taint analysis is not tainting the LHS when these expressions are given. The corresponding tests are temporarily marked as ignored in the TaintAnalysisTest.java class.