Remove default clauses of switch statements that are redundant
This is for cases where a switch statement for an enum exists,
and all enum constants are handled in specific cases.
Removing the default clause will make it a compilation warning
(from Eclipse or Google Error Prone)
instead of a runtime error if the respective enum is extended.

Commit automatically generated with the UnnecessaryDefaultInEnumSwitch
check of Google Error Prone (which is not enabled by default).
Unfortunately we cannot enable this check in CI
because it also produces suggestions that result in compilation errors:
https://github.com/google/error-prone/issues/5045

Part of #1339.