Refactor and enhance taint analysis handling logic.
Improved handling of various initializer types (method handleDeclarationEdge()), added a new method for taint checks, and streamlined access to tainted variables. Extended support for additional CExpression types and refined logging for unhandled cases to improve debugging. Most of these log lines are temporary, but they will stay there till the analysis is more complete.

Initializations like `struct Triple t = {x, y, z};` and `int d[3] = {x, y, z};` are now supported and directly tainted when a tainted variable is present on the RHS. This was achieved by implementing the case when the initializer is a CInitializerList.

Checks like:
    __VERIFIER_is_public(e[0], 0);
    __VERIFIER_is_public(d[1], 0);
    __VERIFIER_is_public((char) y, 0);
    __VERIFIER_is_public((char) (x + y), 0);
    __VERIFIER_is_public(t, 0);
    __VERIFIER_is_public(&t, 0);
    __VERIFIER_is_public(t.a, 0);
    __VERIFIER_is_public(t.b, 0);
    __VERIFIER_is_public((int[]){x, y, z}, 0);

are now as well directly supported as the first argument in the __VERIFIER_is_public() check. This was achieved in the handleStatementEdge method.