Refactor taint analysis parameter naming and move handling of the __VERIFIER_is_public case.
Renamed `firstArg` to `exprToCheck` for clarity and consistency.
The lines

```
Set<CIdExpression> exprToCheckParams = TaintAnalysisUtils.getAllVarsAsCExpr(exprToCheck);
```

and

```
expressionIsTainted = taintedVariables.stream().anyMatch(exprToCheckParams::contains);
```

are handling the publicity check, relying on the CollectCIdExpressionsVisitor class. The `instanceof` checks are still there, but the functionality was removed from the most of them. The goal of leaving them there is to be able to observe the behavior of the benchmark programs.