Refactor taintBySqrtSafe.c
Updated variable declarations and initialization. This is not directly related to the SQRT function but shows that the taint analysis supports different kinds of declarations and initializations.

Adjusted extern function signatures; renamed the parameter of SQRT, because unnecessary1 at the end, and updated the return value of the public-state check to minimize the terminal analysis logs.

Regarding the test logic: now the expression SQRT(x*y) is directly passed to the variable z, since with the last updates of the analysis code, now it is not necessary to save the inner multiplication in another variable for the taint analysis to be able to handle the information flow of the statement correctly.

Added one line at the end to show that similarly to the last explanation, the public-state check supports now more kind of expressions, like e.g., in this case extern function calls as first argument. (Previously only variables like `x` were supported there).