Remove "manual" string parsing for unknown RHS.
Eliminated the `rhsOfRawStatementContainsTaintedParameters` method and associated logic for analyzing raw RHS expressions.

The removed functionality handled cases like calls of `sizeof` or `x && y`. Its functionality is not unnecessary for accurate taint detection, but the "local" parsing of the raw string is not a suitable solution for maintainability and proper code structure, since the responsibility of the analysis is not to parse the edges, but analyse the information flow.

The reason for the method `rhsOfRawStatementContainsTaintedParameters` to exist was the fact that CPAchecker evaluate certain expressions before passing them to the taint analysis. For that reason, the taint analysis can not track the taint flow from the RHS to the LHS properly for such expressions, because the RHS was already evaluated at that point, overriding the information about the RHS contained values.
The only information about the RHS variables is contained in the (string) raw statement, which the method rhsOfRawStatementContainsTaintedParameters was parsing to be able to track the taint flow correctly. However, the parsing implemented in that method was incomplete and letting several cases unhandled, therefore, and to complain with the separation of concerns, it was decided to remove it.

The current workaround for handling function calls like `sizeof` is to use extern defined functions, e.g., `sizeOf`. This, although not totally ideal, and considering that the actual functionality of the function is not important for the information flow, at least allows the analysis to track the taint flow correctly.
This fix partially the case of function calls like `sizeof`, but the case for operators like && or || stays unhandled.

Given this problem there are three possible approaches:

1. to let every function call assignment, for which CPAchecker delivers an evaluated expression instead of a parsed one, to taint the respective LHS.
Problem with solution 1: There would be no difference in the handling of assignments like `int x = sizeof(y)` and `int x = 1`, resulting in the last one tainting its LHS. This would be wrong.

2. Assuming that all the functions like `sizeof` do not really propagate sensitive information, to let every function call assignment, for which CPAchecker delivers an evaluated expression instead of a parsed one, not to taint the respective LHS.
Problem with solution 2: We are not certain that every function that will be evaluated in this case will not be a source of sensitive information. Additionally we would break the basic concept of the taint analysis overapproximation, that every RHS containing a tainted variable, taints its respective LHS.

3. Ideal Solution: for such function calls assignments like `x = sizeof(y)` or the logical operator `&&` CPAchecker not to only deliver the rawStatement, but also the parsed original variables or expressions contained in the RHS of the original edge.

Resuming: The current workaround for this, is to use external function calls instead of calls like `sizeof`, and for && and || to let them not be source of tainted information, inclusive when a tainted variable is contained in the comparison. E.g., given a tainted `y`, the assignment `int x = y && z` will not taint x. (which is still a violation of the mentioned overapproximation).