Refactor taint analysis transfer logic for function calls and return edges
Replaces the simple state cloning with detailed handling of function call arguments and parameters. The new implementation tracks variable tainting, determines killed/generated variables, and evaluates assignment expressions. This change enhances taint precision.

An example of what this addition does can be seen in the line
`__VERIFIER_is_public(f(a, b), 0);`
assuming a function f(int x, int y), a tainted, and b untainted.

Now the analysis will temporarily make the taint flow from tainted entered parameters to the function arguments. This taint will be present only during the scope of the called function. The function variables are going to be untainted after returning the value of the called function. Now the information flow of more complex expressions like

`__VERIFIER_is_public(f(a + 2, f(a * b, 1)), 0);`

is properly tracked. Before, the analysis was only checking whether a tainted variable is present, and if so, returning a taint flow into the called check. Now, the taint will only flow to the returned value when it actually flows into the return statement.