Implement evaluation of assumptions
Till now the analysis was wrongly generating new states for unreachable branches when a condition involving CIdExpression was present in the C programs.

This is now fixed in the handleAssumption method, where the condition is evaluated and creating new states only for reachable branches, and for both only in case we cannot determine the truth value of a condition. For unreachable branches no state is created.

In the theory the assumption can be of any type implementing a CExpression, but I am not certain whether other expressions aside binary can reach the handleAssumption. This still needs to be investigated and expanded if necessary.