Refine isLessOrEqual logic in TaintAnalysisState.
Updated the method to perform a detailed comparison of tainted and untainted variables, ensuring both sets match in keys and values. This enhances the accuracy of state comparisons and fixes potential inconsistencies that could lead, for example, to a premature stop of the analysis execution.

One state is "less or equal" than other state when the other contains all the tainted and untainted variables that this state contains, with each variable having the same value. So this check will fail in two cases:
1. The other state does not contain all the variables of this state.
2. The other state contains all the variables that this state contains, but there is at least one variable that has a different value that the one in this state.