Refactor and simplify taint analysis for the public-state check logic
Removed type-based logging and streamlined `checkInformationFlowViolation` handling.

The changes to the method `checkInformationFlowViolation` will be made in the next commit to keep this commit not so messy.

The main changes here involve the handling of conditional branches. The analysis was dependent of the order in which the conditions were evaluated, e.g., for tainted `x` and untainted `y` the calls:

__VERIFIER_is_public(x < 0 ? y : x, 0);

and

__VERIFIER_is_public(x < 0 ? x : y, 0);

produced different results when the condition could not be evaluated. Now the order of the analysed branches in this scenario is not relevant anymore.

The logic achieved with these changes can be resumed as follows:
The analysis goes through both branches, identifying property violations, but not stopping until the joined state, in which both branches come together, was analyzed. If one of the branches tainted a variable while the other not, the variable will be considered as tainted as per t(z) = t(x) + t (y) = T + U = T in the example above, with t a function that returns the taint state of a variable.