Refactor taint handling in parameter processing.
Removed unnecessary TODO comment and streamlined variable handling by adding function parameters to `killedVars`. The previous approach for this was to completely delete the local variables. The problem is that we might need some calculations including these values in the state generation. Now, instead we preserve these variables, so, they will live after their scope. But this does not represent a problem because the equality for these variables is not based on raw names, but in unique properties that CIdExpressions have (qualified name, file location etc.).