Refactor variable lookup to use recursive call to evaluateExpression for CIdExpressions.
This was made because the implementation of pointer handling forced the analysis to be able to not only return CLiteralExpressions as evaluated values, but more generally CExpressions, to be able to map tainted and untainted variables to CIdExpressions.
Now the result of the changed lines is not always going to be a CLiteralExpression, it could be a CIDExpression as well. To mantain the implementation consisten we change the direct look-up call with a recursive call that will always handle the right data type.