Remove redundant cleanup of function parameter taints
Eliminated unnecessary code that removed taint information for function parameters after a function call. The goal of this check was to "clean" the local variables defined in a function outside the mine function from the tainted and untainted maps after leaving the function.
This is not necessary because the presence of these values does not affect the information-flow analysis. The situation I wanted to avoid was that the maps contain several variables with the same name. After some investigation it was clear to me that duplicated names in the maps do not represent a problem, because of the robust definition of CExpressions in CPAchecker. The presence of these variables in the maps of tainted and untainted variables become garbage in worst case. In case a function is called several times, the local variables will be replaced and updated with the new context.