Reduce `MAX_ALLOWED_STATE_SUCCESSORS` to 50 and adjust loop bounds in corresponding taint analysis benchmark tests.
This aims to fix the Unit Tests in the CI that are failing due to timeout.

The MAX_ALLOWED_STATE_SUCCESSORS is meant to be deleted eventually and is there for testing purposes.
The adjusted benchmark has now a smaller number in the conditions, due to the reduction of the allowed successor states.