add thread active counter when using NUM_STATEMENTS as NonDet source
the counter increases for each created thread and decreases with every terminated thread. initialized with 1, it reaches 0 if no thread is active anymore. this was previously done with the next_thread variable, which is not used in the NUM_STATEMENTS configuration. the non-deterministic next_thread (probably) introduced an enormous overhead