Improve the matching of the file locations inside of CPAchecker
This fixes a bug where the import and export of the precision witness for the file ../../sv-benchmarks/c/loop-simple/nested_4.c was not working correctly. This was caused by a lacking match in the CFA for the AST element and the match not being exact, since the loop head is before an expression evaluation i.e. the condition of the for loop instead of at the beginning of a statement i.e. the for loop