Disable LLVM frontend and mark it as unmaintained
CPAchecker now only prints an error message if the LLVM frontend is
used, no matter whether a command-line parameter or the configuration
option is used or the LLVM IR language is auto-detected as input.

The code is also marked as unmaintained and excluded from some static
analyzers.

The llvm-j library is kept as a dependency because otherwise our code
would no longer compile. It is also quite small.
But we exclude the large LLVM binary library.
This saves 70 MB, 1/6 of the total size of CPAchecker.

Fixes #1356