Prefer Java 21 when installing .deb package (if no Java is installed)
Because it is often faster, we prefer users to use Java 21
to run CPAchecker instead of Java 17 (cf. previous commit).

With the changes from this commit, apt will resolve the Java dependency
of CPAchecker's .deb package as follows:
- If Java 17 or newer is already installed, the dependency is satisfied
  and no additional Java package will be installed.
- Otherwise, if OpenJDK 21 is available, install it
  (e.g., Ubuntu 22.04 or newer, Debian 13 or newer).
- Otherwise, if OpenJDK 17 is available, install it
  (e.g., Debian 12).
- Otherwise the installation fails (we need Java 17 or newer).

This will become effective for the next build of CPAchecker's .deb package,
i.e., for the next CPAchecker release.