Use type SequencedMap/Set for LinkedHashMap/Sets
Presumably, if one creates a LinkedHashMap/Set, the iteration order is
relevant, otherwise one would use a plain HashMap/Set.
In this case, it is good to expose this information in the type,
which since Java 21 is possible with the SequencedMap/Set types.
This allows for example to quickly see whether some iteration over a set
happens to use a set with undefined order or a defined order.

Also add documentation explaining this.

This commit is just a first step and changes types only where something
is initialized with LinkedHashMap/Set. If we continue to strengthen the
types in the future, we can maybe eventually have static checks that we
do not iterate over sets with undefined order.