Simplify `isLessOrEqual` checks in `TaintAnalysisState` by using key set containment instead of entry matching
This makes the isLessOrEqual method to ignore the current assigned values for the check. This decouples the relation between program variables and their current values, since the TaintAnalysis CPA should not be tracking the values. The tracking of values is intended to be removed from TaintAnalysis, because the predicate analysis should take care of checking the feasibility of the encountered violations. Although, the values are still being used in the analysis for checking reachability, because I have not encountered the way to make the predicate analysis to make these checks correctly.