Use SequencedCollection.getFirst intead of List.get(0)
In this case it is somewhat unclear whether we should do it.
List.get(0) is shorter and has no actual disadvantage,
it is just about readability.
I tend to prefer getFirst because it uses wording instead of magic
numbers. The less numbers appear in the source in general,
the more the remaining numbers stand out.
It is also more consistent with getLast(),
which of course is significantly better than its get(int) alternative.

Commit automatically generated with Refaster.