Use SequencedCollection.addFirst and removeFirst instead of add/remove
This is nicer in most cases (except where we use for example add(0) in
combination with add(1), but these are rare and excluded here).

Note that all the cases touched in this commit have bad performance
if the used list is an ArrayList, and it seems this is actually the case
in a surprisingly high number of cases.
Every add(0) and remove(0) operation on an ArrayList involves moving
each entry and is O(n)!
So these should probably be refactored to use ArrayDeque,
or modify the ArrayList only at the end, not the beginning.
But this is orthogonal to the current change.

Commit automatically generated with Refaster.