Refactor `getStatesWithSingeValueMapping` in `TaintAnalysisUtils` to improve state handling, reduce redundancy, and enhance clarity in taint updates.
Merged states that did not result in a full join, but only a partial one, are a sort of state-containers that carry the tainted status of all variables in the states, but with more than one value mapped for each contained variable. In this part of the process, the original states that are being carried by this state-container are reconstructed but with updated taint.

Once more, this solution is not 100% corresponding to the taint analysis concepts, but since the path-sensitivity demands value tracking for the variables, I've implemented this.
Ideally, the functionality for tracking variable values should be made in another CPA, but I did not have time for testing a new implementation with separated concerns.

The modified methods ensure now that each method that was contained in the state-container will be handled and the taint will be propagated correctly for each one.