Refactoring: use pattern-matching switch for cases with maybe null
This is the biggest new feature of Java 21
and allows us to switch over types as done in this commit.
No pattern matching with deconstruction of records is done yet.

This commit changes only cases where I am not sure whether null needs to
be handled and keeps the behavior of the code the same as before
(grouping null together with the default case) and adds a TODO.

Commit based on refactoring by Eclipse with some manual adjustments.