Refactoring: Move withQualifiersSetTo into CType
This allows us to use dynamic binding for selecting the right method
instead of having to use a switch over all subtypes of CType.
The disadvantage is that we cannot make the return type generic and
guarantee to the caller that they get the same type as they pass in
(at least unless making CType generic).
But almost no code makes use of that anyway.