Store qualifiers of a CType in a separate CTypeQualifiers object
So far we use two booleans to store the const and volatile qualifiers.
But this means that there are many methods with two boolean parameters,
and it is easy to mix them up.
Because we need to add more qualifiers in the future (_Atomic),
and the situation would be even worse then,
we need to refactor this and introduce a separate object for qualifiers
that we can pass around as a single parameter.

This commit is the first part of this and adds the new CTypeQualifiers
class and uses it mostly internally in CType subclasses.
The old constructors of those subclasses (with two boolean parameters)
still exist because they are used by many other parts of CPAchecker.
This will be refactored in a separate commit.

For CTypeQualifiers we only need four instances, and eight in the future
(with _Atomic). So we can use a multiton and pre-create these instances.
In Java typically the best way to do this is to use an enum,
even though we do not expect that it will be used in a switch.

An alternative to the multiton would be a record,
which would allow using pattern matching.
But so far we don't need this.