Simplify some qualifier handling in ASTTypeConverter
The previous code merges information about qualifiers from two sources
(taking the union of present qualifiers).
In one of the cases this is actually not needed and we can replace it
with a check that one of the sources has no qualifiers.
In the other case we can use a new union method for CTypeQualifiers,
such that less code outside of CTypeQualifiers has to know which
qualifiers exist.