Update Refaster rules and remove unnecessary Collections.unmodifiable*
If we already know that some collection is immutable,
we do not need to call one of the Collections.unmodifiable* methods.
Directly passing the input collection is better, because it allows us to
benefit from the (stronger) type of the input,
such as if we know it is an ImmutableCollection.
For example, ImmutableSet.copyOf(Collections.unmodifiableSet(coll))
always copies coll, but ImmutableSet.copyOf(coll) might not need to.

Commit automatically generated with Refaster plus some manual type
strengthenings of return types afterwards.