Rename CTypeQualifiers.isConst/Volatile to containsConst/Volatile
While for CType, the naming "isConst" was and is fine
(a type can be said to be const),
for CTypeQualifiers, which represents a set of qualifiers,
this naming can be confusing:
"is const" could mean either "equals CTypeQualifiers.CONST" or
"contains const", and the current semantics are the latter.
So we rename the methods and fields to make this clear.