Fix conversion of typedef types if Eclipse doesn't resolve them
If we convert a typedef type, we want to use a CTypedefType,
and not resolve the typedef to the underlying type.
This was done when converting ITypedef instances but not when converting
IASTNamedTypeSpecifier instances. Now do it there as well.

This was triggered by a new sanity check introduced recently,
which produced some exception when typedef types occurred in casts in
witnesses (where the Eclipse type lookup does not work).