Check that every instance of _Atomic is actually handled
There are some cases, like "int * _Atomic a;", where it seems we cannot
see the _Atomic in the AST, even after applying around workaround of
replacing _Atomic with an attribute, because the attributes for
declarators and pointer operators are always missing.
So CPAchecker would silently treat the above as "int *a;".
In order to avoid this, this commit adds a sanity check.
We collect all instances where our workaround is applied and _Atomic was
seen by the preprocessor, and then check after parsing if we handled all
of then. If not, we produce a parsing error.