Fix caching in CachingTargetLocationProvider
Since 89bb693c the caching did not work anymore,
on every request the target locations were recomputed.
The reason is that with weakKeys() the cache uses identity comparisons
and not equals(), but we use a new cache-key instance on every lookup.

But weakKeys() and weakValues() seem useless here anyway.
We only ever use the target-location provider with the main
specification of the analysis and the specification for finding loop
heads, so there are at most two keys.
And the result is a set of CFANodes that is typically small.
So we can just keep this in memory as long as the analysis runs.

Fixes #1381