Small simplification of CFAUtils.all*Edges
We can pull the check whether the node has summary edges up
out of the FluentIterable, and just delegate to
enteringEdges/leavingEdges if there are no summary edges.
This changes behavior only if a summary edge is added between the call
to the method and the use of the iterable, but CFANodes should remain
unchanged during the analysis.

This change is a preparation for future optimizations,
where we can directly return some collection from
enteringEdges/leavingEdges instead of an iterable.