Refactoring: Move CFAUtils.*edges to CFANode classes
The old methods are kept in place for now and delegate.
No callers are changed.

The methods behave the same and do not yet make use of any optimizations.
The only difference is that some of them threw AssertionError before
and now they return a real value, because due to dynamic binding they
may be called in legitimate cases, whereas previously with static
binding and overload resolution they could be called only in useless cases.

The naming is adjusted to match the other CFANode methods
(include the prefix "get"), and some of the methods that target specific
CFANode subclasses need to be renamed due to overloading restrictions.

For one of the latter methods the return type is actually wrong (cf. #1319).
This commit does not change that.