Deprecate and inline CFAUtils.*edges methods
Now we can directly use the new methods on CFANode.

Contrary to our typical development process, the old methods in CFAUtils
are left for now and not deleted.
They do not hurt and this likely makes merging old branches easier.

The inlining in this commit was done automatically
with Google Error Prone.