Optimize CFANode.get*Edges().contains() for extreme cases
Typically, CFANodes have only 1-2 entering and leaving edges.
There it does not matter what we use as a data structure,
even an O(n) contains check is fast if n==2
(using a set would likely not be faster).

However, there are two cases where the number of entering and leaving
edges is basically unbounded: entering edges on FunctionEntryNodes and
leaving edges on FunctionExitNodes.
And it turns out that because LocationTransferRelation does
getAllLeavingEdges().contains(), this can be a huge performance problem
(cf. #1388).
So we add an optimization for these two cases and use a set.

This commit implements the minimal solution and changes nothing else.
In particular, it does not change the data structure used for storing
entering and leaving edges of CFANodes in general.
This is safer, because maybe sometimes we have edges that are treated as
duplicates due to their equals() implementation and would be lost
if we change to using a set. Maybe this case does not exist,
but it would at least need further investigation.

For the two cases mentioned above, we are sure that there are no problems.
We also do not need to take care of getAllEnteringEdges()/getAllLeavingEdges(),
because we know that for the two relevant cases there are no summary
edges and thus the getAll*Edges() methods just delegate to the variants
without "all" in the name.

We can still wrap the set in a FluentIterable, because whenever possible
the FluentIterable delegates calls to the wrapped data structure if the
latter is a Collection.

So taken together, although it is not visible when looking at the types
of the CFANode API and it might be a little bit brittle because it
depends on a few pieces of code all fitting together in the right way,
it does achieve the performance improvement
that is necessary to fix #1388.

Of course we can in the future revisit this and maybe improve the
data structures of the CFA more generally, but for now this is not needed.