Fix case in predicate analysis were wrong equals was used for floats
For floats in SMT there is the regular equality but also the one that
encodes IEEE754 semantics, i.e., where two NaN values are never equal to
each other.
In the code here we do not perform an equality comparison,
but copy a value from one place to another.
So we need the regular equality.

Fixes #1166.
Thanks Daniel Raffler for the help in the investigation.