[AstCfaRelation] Improve the heuristics to find the loophead for an iteration statement
Currently there was a problem matching this for the case of `while(1) {
... }`. Since the condition of the loop was empty, since it was removed
by our frontend parser.

One file where this occurs is: ../../sv-benchmarks/c/product-lines/elevator_spec1_product01.cil.c

A more permanent solution, not based on a set of heuristics, is still needed, but this likely requires a
complete re-work of the AstCfaRelation.