Precise width scaling of Report.html popovers
Popovers (view of value assignments in counterexample view, help texts)
are limited by our bootstrap version to max-width: 276px.
This is too narrow for useful help texts. So I assume that to circumvent this,
a higher minimum width of 500px was set in the past. But now popovers that do not
require that much space are unnecessarily wide.

To fix the issue correctly, override bootstrap's default max width for popovers.