Use Ruff as code formatter for Python code instead of Black
Ruff is a newer tool that shares the same spirit
and has the goal of producing basically the same format as Black.
There are only minimal differences, but where there are,
Ruff is slightly nicer (cf. diff of this commit).
It is also faster and we will also use it as flake8 replacement.

While we are at it, raise the minimum Python version that the formatter
should use to 3.10 (this has no actual effect on our current code).

The CI integration is based on
https://docs.astral.sh/ruff/integrations/#gitlab-cicd