Refactor resolution of field designators into external method
Do not modify outer list from within switch-expression, but handle resolution of field designator
as special case that may produce multiple internal CFieldDesignators.
Because this does not match the type of the switch-expression anymore (single CDesignator),
do not force everything to be a List<CDesignator>, but extract the resolution of field designators
from the switch-expression into a separate branch.