Remove duplication in handling of Abstract states with assumptions
Currently for each state, we get all its substates, then compute all
possible substates each of them may have, and get all assumptions out of
all of them.

This leads to a duplication of many states, which are then added to the
PathFormula for the SMT-Solver.
I don't suspect this has a huge performance difference, but is
conceptually sounder, and also aids in debugging, since it keeps the
path-formulas smaller