Weaken condition on type passed to conversion of array designators
If the designator is a pure array, then the conversion
of InitializerList will already resolve the array type.
Two examples where this happens:
```
int arr[2] = {5, 6}
int arr[2] = { [0] = 5, [1] = 6 }
```

The type used to resolve initializer `{5, 6}`
and initializer `{[0] = 5, [1] = 6}` should be the same,
and it should be the type of the array's elements (here: int).
This requires a resolution from the type the initializer assigns
to (here: int[]).

This resolution can not be done in the resolution of designators only,
because `{5, 6}` has no designators.

The resolution can also not be only done in the resolution
of an initializer list (`{..}`),
because this would miss the resolution for a designator like
`.a[0].x`.

So for now, we resolve array types
both when an array designator is resolved and when
an initializer list is handled.