Normalize the path before a lookup in the CSourceOriginMapping making sure that every path is interpreted the same
Current for #1469 the problem was that a lookup of the line in the original program in the `CSourceOriginMapping`
provided the provided line number, since no code delta was found.
This was caused by the path being used for the lookup being semantically
equivalent to the one provided during the creation of the mapping, but
them differing syntactically.

Only the witness version 2.0 export was affected since they re-analyse
the AST and during this re-analysis apparently a slightly different
file-name was used.