Fix missing default resource limits in most important configs
For many years CPAchecker wanted to have and had default resource limits
in its main configs.
Unfortunately this was removed for the default config in 1935715d.
For SV-COMP, we had a required time limit so users were forced to
specify the time limit on the command line.
But when we removed this requirement in 65ba664a,
we forgot to add the default time limit.

This commit adds the default resource limits back for the default and
the current SV-COMP config.