fix issue with 'analysis.summaryEdges'
the problem was that CFunctionSummaryStatementEdges altered the CFA in such a way which resulted in a few checks to crash. this commit essentially excludes all additional CFANodes + CFAEdges that are introduced by this option. this also highlighted another issue, with the non-inlining of reach_errors. that is now simplified and not handled in any special case, except for an injected call to reach_error when assigning parameters, so reach_errors are now handled (almost) the same as any other function.