Fix preprocessing for wrong architecture (32/64 bit) on x86
When CPAchecker calls a preprocessor (cpp or clang),
it needs to pass the target architecture such that the preprocessor
can choose the right headers.
This is especially important because CPAchecker defaults to 32 bit
and the typical preprocessors default to the current host architecture,
which is mostly 64 bit.
So except when passing --64 to CPAchecker the preprocessing
was typically wrong.

This commit fixes the behavior only for x86.
For ARM the compilers treat 32 bit and 64 bit as different
architectures, not variants of the same architecture,
and I am not sure which argument we would need to pass.
For the same reason, preprocessing of a task for a completely different
architecture (ARM tasks on x86 or vice versa) is also still wrong.