Move place where Configuration.dumpUsedOptionsTo is called
So far we did it immediately after switching configs
based on language, witness, and property.
Now we do it a little bit later when we dump the config to a file.
This means that the options for creating the logger and those in
MainOptions will not be dumped with --print-used-options,
but this change makes the refactorings in the next commits much easier
because there are fewer stages of option injection.
The feature --print-used-options is rarely used anyway.