Simplify method for language detection based on passed files
So far it stored the detected language in the BootstrapLanguageOptions
object and returned a new Configuration instance, also with the language set.
But the latter was redundant because handleFrontendLanguageOptions()
also takes care of it.
So we can remove all side effects from the detection method and simply
return the language.