Change when logger is instantiated
So far we instantiated it after switching configs based on language,
witness, and property, and the log options were taken from the final
config file.
Now we instantiate it before switching configs and the log options are
taken from the initial config file, differently from all other options
in CPAchecker.
This can be seen as a disadvantage due to inconsistency,
but it has the advantage that the logger is available earlier
and that logging about configuration-injection problems also starts earlier.
Furthermore, logging options should basically never be set
in our config files anyway, so I find this acceptable.