Solve warnings about unused options for bootstrap options
From the point of view of the main Configuration instance that is used
for CPAchecker, all the "bootstrap options" (for switching configs based
on property etc.) are unused, so they trigger warnings.
Furthermore, these options end up in UsedConfiguration.properties,
but would cause problems if one wants to use
UsedConfiguration.properties as config file for another CPAchecker run.

So far both problems were attempted to be solved by clearing some of the
options from the produced Configuration instances after they are used.
But this was done in an ad-hoc manner and was incomplete.
In particular, any option given on the command line would be copied into
all the Configuration instances, so for example if we switch twice
(based on witness and property) then the clearing of config options
after the first switch could have been overwritten by re-adding the
options from the command line after the second switch.

The current solution is also not so nice because it involves some
redundant injection and we still need to manually list config options,
but at least it solves the problem completely.
Because we perform the cleanup of bootstrap options
on the final Configuration instance and do it consistently,
all bootstrap options will not be shown as unused.