Match Acsl Comment to Cfa Node
Currently finds the Cfa Node for an Acsl Comment in twor steps. This
process should be expanded in the future.

- Step 1: Find the Cfa Node by getting the tightest statement for the
comment location from the AstCfaRelation and then subtracting the set of
successors of all edges from the set of predecessors of all edges. This
should return exactly one Cfa Node but does not work for function
contracts because for those the tightest statement is null.

- Step 2: Check if the remaining comments come directly before a
function entry node. In this case, they are fynction contracts.

At the end there should not remain any Acsl Comments without a Cfa Node.

- Known Bug: In some cases, we get more than one Cfa Node during the
first step. This is because some nodes falsely don't occur in the
successor set.