Get the statement for a location
- Introduces the private method getStartingFor(). This method takes as
input a line number and an optional column number of a file location. It
returns an Optional Entry of the corresponding StartingLocation and its
AstElement.

- The public method getTightestStatingForStarting() calls the private
method getStartingFor(). If getStartingFor() has a result and the line
number of the results starting number matches the line number of the
caller file location, this method returns the AstElement that was
returned by getStartingFor(). Otherwise, it returns an empty optional.

This method should be used to get the tightest statement for program
code.

- The public method getElemForStarting() also calls getStartingFor() and
passes on the AstElement if it exists without comparing the line
numbers. If getStartingFor() has now result, getElemForStarting()
returns an empty optional

This method should be used to get the tightest statement for an
annotation/comment.