Sort Function Entry Nodes by File Location
To identify function contracts from the acsl annotations we create a
sorted set of all function entry nodes and then look for the one with
the smallest file location that is still greater than the location of
the acsl annotation.

In the next step we need to check that there is no other node inbetween
the annotaton and the function entry node.